299 run.sh refactor

+++ cppProj/2021Fenix/run.sh
@@ -1,11 +1,16 @@
 executable=a.out
-{
-  rm -v $executable # in unix, a.exe would be a.out
-  g++ -I. -std=c++17 main.cpp -o $executable
+function1(){ # a function to support early return
+  [[ -e $executable ]] && rm -v $executable 
+  g++ -I. -std=c++17 main.cpp -o $executable || return -1
   #g++ -I. -std=c++0x board.o Engine1.o Engine2.o main.cpp  
   printf "\n----- build done, now running -----\n" 
-  # cat stdin.txt |./a.exe
+  # cat stdin.txt |./$executable
   ./$executable
-} > log.txt 2>&1
-cat log.txt
-# Based on TickDB run.sh + concretize run.sh
+}
+
+function1 > log.txt 2>&1
+
+# good for brief output .. always shows on console, but no pagination for long output
+cat log.txt 
+less -E log.txt 2>&1 
+# good for long output but shows nothing on console for short output
